chore: fix smells
